+++ pkg/adapter/restful/gin/carsrs/resource.go
@@ -18,30 +18,31 @@ import (
 )
 
 type resource struct {
-	cars *carsuc.UseCase
+	cars func() *carsuc.UseCase
 }
 
 // Register instantiates a resource adapting the cars use case instance
 // with the relevant REST APIs including:
 //  1. PATCH request to /api/caweb/v1/cars/:cid
 //     in order to ride or park a car.
-func Register(r *gin.RouterGroup, cars *carsuc.UseCase) {
+func Register(r *gin.RouterGroup, cars func() *carsuc.UseCase) {
 	rs := &resource{cars: cars}
 	r.PATCH("cars/:cid", rs.UpdateCar)
 }
 
 func (rs *resource) UpdateCar(c *gin.Context) {
-	req := rs.DserUpdateCarReq(c)
-	if req == nil {
+	req, ok := rs.DserUpdateCarReq(c)
+	if !ok {
 		return
 	}
+	carsUseCase := rs.cars()
 	var car *model.Car
 	var err error
 	switch req.Op {
 	case "ride":
-		car, err = rs.cars.Ride(c, req.CarID, req.Dst)
+		car, err = carsUseCase.Ride(c, req.CarID, req.Dst)
 	case "park":
-		car, err = rs.cars.Park(c, req.CarID, req.Mode)
+		car, err = carsUseCase.Park(c, req.CarID, req.Mode)
 	default:
 		panic("unexpected op:" + req.Op)
 	}