+++ pkg/adapter/restful/gin/gin_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/momeni/clean-arch/internal/test/dbcontainer"
 	"github.com/momeni/clean-arch/pkg/adapter/config/cfg2"
 	"github.com/momeni/clean-arch/pkg/adapter/config/settings"
+	"github.com/momeni/clean-arch/pkg/adapter/config/vers"
 	"github.com/momeni/clean-arch/pkg/adapter/db/postgres"
 	"github.com/momeni/clean-arch/pkg/adapter/restful/gin"
 	"github.com/momeni/clean-arch/pkg/adapter/restful/gin/routes"
@@ -83,11 +84,22 @@ func (igts *IntegrationGinTestSuite) SetupSuite() {
 	igts.Gin = gin.New(gin.Logger(), gin.Recovery())
 	igts.Require().NotNil(igts.Gin, "cannot instantiate Gin engine")
 	delay := settings.Duration(2 * time.Second)
-	err = routes.Register(igts.Gin, igts.Pool, cfg2.Usecases{
-		Cars: cfg2.Cars{
-			DelayOfOPM: &delay,
+	c := &cfg2.Config{
+		Usecases: cfg2.Usecases{
+			Cars: cfg2.Cars{
+				DelayOfOPM: &delay,
+			},
 		},
-	})
+		Vers: vers.Config{
+			Versions: vers.Versions{
+				Database: postgres.Version,
+				Config:   cfg2.Version,
+			},
+		},
+	}
+	err = c.ValidateAndNormalize()
+	igts.Require().NoError(err, "preparing configuration settings")
+	err = routes.Register(igts.Ctx, igts.Gin, igts.Pool, c)
 	igts.Require().NoError(err, "failed to register Gin routes")
 }