+++ pkg/adapter/restful/gin/settingsrs/serdser.go
@@ -1,0 +1,23 @@
+// Copyright (c) 2024 Behnam Momeni
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+package settingsrs
+
+import (
+	"github.com/gin-gonic/gin"
+	"github.com/gin-gonic/gin/binding"
+	"github.com/momeni/clean-arch/pkg/adapter/restful/gin/serdser"
+	"github.com/momeni/clean-arch/pkg/core/model"
+)
+
+func (rs *resource) DserUpdateSettingsReq(
+	c *gin.Context,
+) (*model.Settings, bool) {
+	req := &model.Settings{}
+	if ok := serdser.Bind(c, req, binding.JSON); !ok {
+		return nil, false
+	}
+	return req, true
+}