+++ pkg/core/usecase/appuc/options.go
@@ -1,0 +1,9 @@
+// Copyright (c) 2024 Behnam Momeni
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+package appuc
+
+// Option is a functional option for the application use case.
+type Option func(uc *UseCase) error